separate functionalities into a module
